Move test funcs to utils/test.go
